Rudimentary transposition table

+++ src/engine/search/evaluation.rs
@@ -1,6 +1,6 @@
 use super::*;
 
-pub const MATERIAL_WEIGHTS: [i32; 12] = [100, 300, 350, 500, 1000, 10000, -100, -300, -350, -500, -1000, -10000];
+pub const MATERIAL_WEIGHTS: [i16; 12] = [100, 300, 350, 500, 1000, 10000, -100, -300, -350, -500, -1000, -10000];
 /*pub const STACKED_PAWN_PENALTY: i32 = -10;
 pub const ISOLATED_PAWN_PENALTY: i32 = -10;
 pub const PASSED_WHITE_PAWN_BONUS: [i32; 8] = [ 0, 10, 30, 50, 75, 100, 150, 200 ];
@@ -10,7 +10,7 @@ pub const OPEN_FILE_SCORE: i32 = 15;
 pub const PROTECTED_KING_BONUS: i32 = 5;*/
 
 /// Pawn positional score
-pub const PAWN_SCORES: [i32; 64] = 
+pub const PAWN_SCORES: [i16; 64] = 
 [
     90,  90,  90,  90,  90,  90,  90,  90,
     30,  30,  30,  40,  40,  30,  30,  30,
@@ -23,7 +23,7 @@ pub const PAWN_SCORES: [i32; 64] =
 ];
 
 /// Knight positional score
-pub const KNIGHT_SCORES: [i32; 64] = 
+pub const KNIGHT_SCORES: [i16; 64] = 
 [
     -5,   0,   0,   0,   0,   0,   0,  -5,
     -5,   0,   0,  10,  10,   0,   0,  -5,
@@ -36,7 +36,7 @@ pub const KNIGHT_SCORES: [i32; 64] =
 ];
 
 /// Bishop positional score
-pub const BISHOP_SCORES: [i32; 64] = 
+pub const BISHOP_SCORES: [i16; 64] = 
 [
      0,   0,   0,   0,   0,   0,   0,   0,
      0,   0,   0,   0,   0,   0,   0,   0,
@@ -50,7 +50,7 @@ pub const BISHOP_SCORES: [i32; 64] =
 ];
 
 /// Rook positional score
-pub const ROOK_SCORES: [i32; 64] = 
+pub const ROOK_SCORES: [i16; 64] = 
 [
     50,  50,  50,  50,  50,  50,  50,  50,
     50,  50,  50,  50,  50,  50,  50,  50,
@@ -64,7 +64,7 @@ pub const ROOK_SCORES: [i32; 64] =
 ];
 
 /// King positional score
-pub const KING_SCORES: [i32; 64] = 
+pub const KING_SCORES: [i16; 64] = 
 [
      0,   0,   0,   0,   0,   0,   0,   0,
      0,   0,   5,   5,   5,   5,   0,   0,
@@ -90,8 +90,8 @@ pub const MIRRORED: [usize; 64] =
 ];
 
 impl Position {
-    pub fn evaluate(&self) -> i32 {
-        let mut score: i32 = 0;
+    pub fn evaluate(&self) -> i16 {
+        let mut score: i16 = 0;
 
         for bb in 0..12 {
             let board = self.bitboards[bb];
